Send Close-Ok after receiving a channel Close from the server.
The spec says that the response to receiving Close must be to send
Close-Ok. When a channel was being closed due to an exception, Pika
was not sending Close-Ok back, which made it impossible to reopen that
channel (see issue #84).

Additionally, if the channel is already going away, the Close-Ok
callback that the connection adds should be removed. Not doing that
results in a warning about a duplicate callback being added.